Changed TypeEssentials.getSuperType*() methods to return List<Class<?>>, not List<Class<? super T>>.
There is was only a small benefit in using "? super T"; when calling the
methods with a solid, non-wildcard Class reference like
Class<Serializable>.

When calling the methods with Class<? extends Serializable>, the generic
footprint of the methods return types became List<Class<? super ?
extends Serializable>>, which is a wildcard with upper and lower bounds,
nested 2 stages deep in the generics of List. The result was that the
compiler was only able to recognize the return type as List<?> anymore,
which made the methods unusable.